refactor: use tanstack query to handle data fetch
TanStack Query provides the ability to cache and fetch data. This is
useful to prevent us from needing to roll our own state management when
it comes to fetching data. We can now invalidate the cache to refresh
the cache or update the query to switch between filtered and most recent
rolls.