feat: add licenses link for dependencies used
This is kind of just done for the sake of adding attributions. The MIT
license is very permissive and the likelihood of getting hit with
something is low, especially since a lot of web applications don't even
provide these publicly. However, it's just kind of a neat thing that we
may want to do. Not necessarily something to be maintained but just to
include attributions.