docs: plan for terraform code
Current priorities are to create an MVP of the roller application. Since
the infrastructure is very unlikely to change majorly, this is on the
backburner. The README are notes for potential future development to
make sure that the project can be stood up through some simple
tutorials with commands and parameters.